requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
langchain_google_genai
google-generativeai
faiss-cpu